move the header area in the base page as they are the same
